refactor: restructure to DDD purist architecture
- Move to single Tracking bounded context
- Use Entity instead of Model for domain entities
- Use singular names for Domain concepts (Entity, ValueObject, Repository)
- Use singular for Application patterns (Command, Query)
- Separate Unit and Integration tests properly
- Update all namespaces to reflect new structure
