+++ src/Tracking/Domain/ValueObject/IpAddress.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yomali\Tracker\Domain\Visit\ValueObjects;
+namespace Yomali\Tracker\Tracking\Domain\ValueObject;
 
 /**
  * IP Address value object
